pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "behavior-os"
version = "0.3.0"
description = "BehaviorOS — compile behavior, not prompts. A behavioral runtime for AI agents."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "anthropic>=0.20",
]

[project.scripts]
mindset = "agentic_mindset.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agentic_mindset"]

[tool.pytest.ini_options]
testpaths = ["tests"]
